fix:Correcion de bugs en Actualizar involucrados psicologo

+++ sicf_Models/Core/SicofaCiudadano.cs
@@ -19,8 +19,8 @@ namespace sicf_Models.Core
         public long? IdDepartamento { get; set; }
         //public long? IdLugarExpedicion { get; set; }
         public int? IdPaisExpedicion { get; set; }
-        public int? IdDepartamentoExpedicion { get; set; }
-        public int? IdMunicipioExpedicion { get; set; }
+        public long? IdDepartamentoExpedicion { get; set; }
+        public long? IdMunicipioExpedicion { get; set; }
         public int? IdPaisNacimiento { get; set; }
         public long? IdCiudMunNacimiento { get; set; }
         public int? IdNivelAcademico { get; set; }
